requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
backcall==0.1.0
blinker==1.4
bson==0.5.8
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dataclasses==0.7
decorator==4.4.1
Django==2.2
django-cors-headers==3.2.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
dnspython==1.16.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
mongoengine==0.19.1
msgpack==0.6.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.5.2
PyHamcrest==2.0.0
pymongo==3.10.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.2.4
traitlets==4.3.3
Twisted==19.10.0
txaio==20.1.1
wcwidth==0.1.8
zope.interface==4.7.1
